requirements.txt
Flask==3.0.3
gunicorn==21.2.0
requests==2.31.0
reportlab==4.0
matplotlib==3.8.0
pillow==10.1.0
